pyproject.toml
[tool.poetry]
name = "djangocon-us-2023-tdd-tutorial"
version = "0.1.0"
description = "This repository has the details about my tutorial on Mastering TDD in Django at DjangoCOn US 2023"
authors = ["Kuldeep Pisda <hello@kdpisda.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
markdown = "^3.5"
django-filter = "^23.3"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.0"}
pyyaml = "^6.0.1"
uritemplate = "^4.1.1"
drf-spectacular = "^0.26.5"
pytest-django = "^4.5.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "server.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py",]
